+++ lib/effects/scale_effect.dart
@@ -3,20 +3,20 @@ import 'package:meta/meta.dart';
 
 import 'dart:math';
 
-import '../vector2f.dart';
+import '../vector2_extension.dart';
 import './effects.dart';
 
 double _direction(double p, double d) => (p - d).sign;
 double _size(double a, double b) => (a - b).abs();
 
 class ScaleEffect extends PositionComponentEffect {
-  Vector2F size;
+  Vector2 size;
   double speed;
   Curve curve;
 
-  Vector2F _original;
-  Vector2F _diff;
-  final Vector2F _dir = Vector2F.zero();
+  Vector2 _original;
+  Vector2 _diff;
+  final Vector2 _dir = Vector2.zero();
 
   ScaleEffect({
     @required this.size,
@@ -35,7 +35,7 @@ class ScaleEffect extends PositionComponentEffect {
     }
 
     _original = component.toSize();
-    _diff = Vector2F(
+    _diff = Vector2(
       _size(_original.x, size.x),
       _size(_original.y, size.y),
     );