+++ lib/game/base_game.dart
@@ -11,7 +11,7 @@ import '../components/component.dart';
 import '../components/mixins/has_game_ref.dart';
 import '../components/mixins/tapable.dart';
 import '../components/position_component.dart';
-import '../vector2f.dart';
+import '../vector2_extension.dart';
 import 'game.dart';
 
 /// This is a more complete and opinionated implementation of Game.
@@ -31,10 +31,10 @@ class BaseGame extends Game with FPSCounter {
   final List<Component> _removeLater = [];
 
   /// Current screen size, updated every resize via the [resize] method hook
-  Vector2F size;
+  Vector2 size;
 
   /// Camera position; every non-HUD component is translated so that the camera position is the top-left corner of the screen.
-  Vector2F camera = Vector2F.zero();
+  Vector2 camera = Vector2.zero();
 
   /// This method is called for every component added, both via [add] and [addLater] methods.
   ///
@@ -138,7 +138,7 @@ class BaseGame extends Game with FPSCounter {
   /// You can override it further to add more custom behaviour, but you should seriously consider calling the super implementation as well.
   @override
   @mustCallSuper
-  void resize(Vector2F size) {
+  void resize(Vector2 size) {
     this.size = size;
     components.forEach((c) => c.resize(size));
   }