+++ test/components/component_test.dart
@@ -2,7 +2,7 @@ import 'dart:ui';
 
 import 'package:flame/components/position_component.dart';
 import 'package:flame/components/sprite_component.dart';
-import 'package:flame/vector2f.dart';
+import 'package:flame/vector2_extension.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -14,7 +14,7 @@ void main() {
       expect(c.toPosition().x, 2.2);
       expect(c.toPosition().y, 3.4);
 
-      c.setByPosition(Vector2F(1.0, 0.0));
+      c.setByPosition(Vector2(1.0, 0.0));
       expect(c.x, 1.0);
       expect(c.y, 0.0);
     });
@@ -26,7 +26,7 @@ void main() {
       expect(c.toSize().x, 2.2);
       expect(c.toSize().y, 3.4);
 
-      c.setBySize(Vector2F(1.0, 0.0));
+      c.setBySize(Vector2(1.0, 0.0));
       expect(c.width, 1.0);
       expect(c.height, 0.0);
     });