+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.0.0
  - Move all box2d related code and examples to the flame_box2d repo
  - Rename Animation to SpriteAnimation
- - Unify all the possible tuples to Vector2
+ - Create extension of Vector2 and unify all tuples to use that class
 
 ## 0.26.0
  - Improving Flame image auto cache